+++ docs_snippets/lib/usages/foundation/collapsible.dart
@@ -1,3 +1,5 @@
+// ignore_for_file: avoid_redundant_argument_values
+
 import 'package:flutter/widgets.dart';
 
 import 'package:forui/forui.dart';
@@ -5,6 +7,7 @@ import 'package:forui/forui.dart';
 const collapsible = FCollapsible(
   // {@category "Core"}
   value: 0.5,
+  axis: .vertical,
   child: Placeholder(),
   // {@endcategory}
 );