+++ forui/test/src/foundation/collapsible/collapsible_golden_test.dart
@@ -6,7 +6,7 @@ import 'package:flutter/material.dart';
 import 'package:flutter_test/flutter_test.dart';
 
 import 'package:forui/forui.dart';
-import '../test_scaffold.dart';
+import '../../test_scaffold.dart';
 
 void main() {
   group('FCollapsible', () {
@@ -25,7 +25,7 @@ void main() {
         await expectLater(find.byType(TestScaffold), matchesGoldenFile('collapsible/${theme.name}/fully-expanded.png'));
       });
 
-      testWidgets('half expanded', (tester) async {
+      testWidgets('half expanded vertical', (tester) async {
         await tester.pumpWidget(
           TestScaffold(
             theme: theme.data,
@@ -36,7 +36,28 @@ void main() {
           ),
         );
 
-        await expectLater(find.byType(TestScaffold), matchesGoldenFile('collapsible/${theme.name}/half-expanded.png'));
+        await expectLater(
+          find.byType(TestScaffold),
+          matchesGoldenFile('collapsible/${theme.name}/half-expanded-vertical.png'),
+        );
+      });
+
+      testWidgets('half expanded horizontal', (tester) async {
+        await tester.pumpWidget(
+          TestScaffold(
+            theme: theme.data,
+            child: const FCollapsible(
+              value: 0.5,
+              axis: Axis.horizontal,
+              child: ColoredBox(color: Colors.yellow, child: SizedBox.square(dimension: 50)),
+            ),
+          ),
+        );
+
+        await expectLater(
+          find.byType(TestScaffold),
+          matchesGoldenFile('collapsible/${theme.name}/half-expanded-horizontal.png'),
+        );
       });
 
       testWidgets('fully collapsed', (tester) async {